requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
urllib3